Empty Pom Animations Finished
Finished implementing all of the animations for Empty Pom
