+++ The Pom Game/Assets/Animations/Empty Pom/Hit 2.anim
@@ -22,15 +22,15 @@ AnimationClip:
     curve:
     - time: 0
       value: {fileID: 2113244508946671555, guid: 6a7d0ce0f56b14049b475b4f6b2fa3c5, type: 3}
-    - time: 0.083333336
+    - time: 0.033333335
       value: {fileID: -8210663478143872627, guid: 6a7d0ce0f56b14049b475b4f6b2fa3c5, type: 3}
-    - time: 0.16666667
+    - time: 0.06666667
       value: {fileID: -5092881230492219737, guid: 6a7d0ce0f56b14049b475b4f6b2fa3c5, type: 3}
-    - time: 0.25
+    - time: 0.1
       value: {fileID: -8471775297385135929, guid: 6a7d0ce0f56b14049b475b4f6b2fa3c5, type: 3}
-    - time: 0.33333334
+    - time: 0.13333334
       value: {fileID: 2111399526936113531, guid: 6a7d0ce0f56b14049b475b4f6b2fa3c5, type: 3}
-    - time: 0.41666666
+    - time: 0.2
       value: {fileID: 2111399526936113531, guid: 6a7d0ce0f56b14049b475b4f6b2fa3c5, type: 3}
     attribute: m_Sprite
     path: Sprite
@@ -65,7 +65,7 @@ AnimationClip:
     m_AdditiveReferencePoseClip: {fileID: 0}
     m_AdditiveReferencePoseTime: 0
     m_StartTime: 0
-    m_StopTime: 0.43333334
+    m_StopTime: 0.21666667
     m_OrientationOffsetY: 0
     m_Level: 0
     m_CycleOffset: 0
