+++ The Pom Game/Assets/Animations/Empty Pom/Hit 3.anim
@@ -22,15 +22,15 @@ AnimationClip:
     curve:
     - time: 0
       value: {fileID: 498796871536486607, guid: 93e8d59e4536dd246ab02a13779572a5, type: 3}
-    - time: 0.083333336
+    - time: 0.033333335
       value: {fileID: -5834937896756814159, guid: 93e8d59e4536dd246ab02a13779572a5, type: 3}
-    - time: 0.16666667
+    - time: 0.06666667
       value: {fileID: 9218793781923225871, guid: 93e8d59e4536dd246ab02a13779572a5, type: 3}
-    - time: 0.25
+    - time: 0.1
       value: {fileID: -3301352202719405775, guid: 93e8d59e4536dd246ab02a13779572a5, type: 3}
-    - time: 0.33333334
+    - time: 0.13333334
       value: {fileID: 1755283027228663769, guid: 93e8d59e4536dd246ab02a13779572a5, type: 3}
-    - time: 0.41666666
+    - time: 0.16666667
       value: {fileID: 1755283027228663769, guid: 93e8d59e4536dd246ab02a13779572a5, type: 3}
     attribute: m_Sprite
     path: Sprite
@@ -65,7 +65,7 @@ AnimationClip:
     m_AdditiveReferencePoseClip: {fileID: 0}
     m_AdditiveReferencePoseTime: 0
     m_StartTime: 0
-    m_StopTime: 0.43333334
+    m_StopTime: 0.18333334
     m_OrientationOffsetY: 0
     m_Level: 0
     m_CycleOffset: 0
