+++ The Pom Game/Assets/Resources/AnimationStates/Death.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b84a70fc5fb78734a8dad7bdb15fd059
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
