+++ The Pom Game/Assets/Resources/AnimationStates/Death/Death.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 90c3f25fc96ae684583946ff8fa7b8a4
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
