+++ The Pom Game/Assets/Resources/AnimationStates/Hit.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 778b22657ba39134a9bdd7be09df0c8e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
