+++ The Pom Game/Assets/Resources/AnimationStates/Hit/Hit.asset
@@ -1,0 +1,19 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 71196214900c82c479db633f3d431a4c, type: 3}
+  m_Name: Hit
+  m_EditorClassIdentifier: Assembly-CSharp::Pom.AnimationHandling.ScriptableObjects.OneShotAnimationState
+  animations:
+  - Hit 1
+  - Hit 2
+  - Hit 3
+  animationTag: Hit
