+++ The Pom Game/Assets/Resources/AnimationStates/Hit/Hit.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ecf7a058aee166d40b018c13b5e2fcde
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
