+++ The Pom Game/Assets/Resources/AnimationStates/Pause.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a8b806c8423efa047a9910404aea97c5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
