+++ The Pom Game/Assets/Resources/AnimationStates/Pause/Pause.asset
@@ -1,0 +1,20 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 43f106d695ba7434385e33f3960803d4, type: 3}
+  m_Name: Pause
+  m_EditorClassIdentifier: Assembly-CSharp::AnimationStateSO
+  animations:
+  - Pause 1
+  - Pause 2
+  - Pause 3
+  - Pause 4
+  - Pause 5
