+++ The Pom Game/Assets/Resources/AnimationStates/Pause/Pause.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d1c97128c80ab7844a433dcf37c0046c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
