+++ The Pom Game/Assets/Scenes/PrototypingLevels/P Sandbox.unity
@@ -188,17 +188,6 @@ PrefabInstance:
     m_AddedGameObjects: []
     m_AddedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: b5dc3af9f8394734e83fbe654ab00c5f, type: 3}
---- !u!114 &37214692 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 7835691178222862130, guid: de96de86e86d12b4d85c9812b64d9650, type: 3}
-  m_PrefabInstance: {fileID: 771282380276041544}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 97280a68168de634f82ef725a67cf2f1, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: Assembly-CSharp::Pom.AnimationHandling.AnimationWrapper
 --- !u!1001 &146052669
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -253,7 +242,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 3441735966342920088, guid: de96de86e86d12b4d85c9812b64d9650, type: 3}
       propertyPath: m_IsActive
-      value: 0
+      value: 1
       objectReference: {fileID: 0}
     m_RemovedComponents: []
     m_RemovedGameObjects: []
@@ -570,7 +559,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 3441735966342920088, guid: de96de86e86d12b4d85c9812b64d9650, type: 3}
       propertyPath: m_IsActive
-      value: 0
+      value: 1
       objectReference: {fileID: 0}
     m_RemovedComponents: []
     m_RemovedGameObjects: []
@@ -1034,7 +1023,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 4495459824194946372, guid: c2e3c88e844d9ae47904888c1a5c78fe, type: 3}
       propertyPath: m_IsActive
-      value: 1
+      value: 0
       objectReference: {fileID: 0}
     m_RemovedComponents: []
     m_RemovedGameObjects: []
@@ -1589,7 +1578,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 3441735966342920088, guid: de96de86e86d12b4d85c9812b64d9650, type: 3}
       propertyPath: m_IsActive
-      value: 0
+      value: 1
       objectReference: {fileID: 0}
     m_RemovedComponents: []
     m_RemovedGameObjects: []
@@ -1724,7 +1713,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 3441735966342920088, guid: de96de86e86d12b4d85c9812b64d9650, type: 3}
       propertyPath: m_IsActive
-      value: 0
+      value: 1
       objectReference: {fileID: 0}
     m_RemovedComponents: []
     m_RemovedGameObjects: []
@@ -1923,7 +1912,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 3441735966342920088, guid: de96de86e86d12b4d85c9812b64d9650, type: 3}
       propertyPath: m_IsActive
-      value: 0
+      value: 1
       objectReference: {fileID: 0}
     m_RemovedComponents: []
     m_RemovedGameObjects: []
@@ -2134,7 +2123,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 3441735966342920088, guid: de96de86e86d12b4d85c9812b64d9650, type: 3}
       propertyPath: m_IsActive
-      value: 0
+      value: 1
       objectReference: {fileID: 0}
     m_RemovedComponents: []
     m_RemovedGameObjects: []
@@ -2348,7 +2337,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 3441735966342920088, guid: de96de86e86d12b4d85c9812b64d9650, type: 3}
       propertyPath: m_IsActive
-      value: 0
+      value: 1
       objectReference: {fileID: 0}
     m_RemovedComponents: []
     m_RemovedGameObjects: []
@@ -2553,26 +2542,6 @@ PrefabInstance:
     - target: {fileID: 3981937185633151470, guid: de96de86e86d12b4d85c9812b64d9650, type: 3}
       propertyPath: m_Layer
       value: 7
-      objectReference: {fileID: 0}
-    - target: {fileID: 7835691178222862130, guid: de96de86e86d12b4d85c9812b64d9650, type: 3}
-      propertyPath: animationStates.Array.size
-      value: 4
-      objectReference: {fileID: 0}
-    - target: {fileID: 7835691178222862130, guid: de96de86e86d12b4d85c9812b64d9650, type: 3}
-      propertyPath: animationStates.Array.data[3].label
-      value: Capture
-      objectReference: {fileID: 0}
-    - target: {fileID: 7835691178222862130, guid: de96de86e86d12b4d85c9812b64d9650, type: 3}
-      propertyPath: animationStates.Array.data[3].stateSO
-      value: 
-      objectReference: {fileID: 11400000, guid: bd69a29004fafd04fa9c063be9999e7f, type: 2}
-    - target: {fileID: 7835691178222862130, guid: de96de86e86d12b4d85c9812b64d9650, type: 3}
-      propertyPath: animationStates.Array.data[3].stateMachine
-      value: 
-      objectReference: {fileID: 37214692}
-    - target: {fileID: 7835691178222862130, guid: de96de86e86d12b4d85c9812b64d9650, type: 3}
-      propertyPath: animationStates.Array.data[3].<Tag>k__BackingField
-      value: 6
       objectReference: {fileID: 0}
     - target: {fileID: 7835691178222862130, guid: de96de86e86d12b4d85c9812b64d9650, type: 3}
       propertyPath: animationMappings.Array.data[2].NestedAnimationSettings.Array.size
