+++ The Pom Game/Assets/Scripts/AnimationHandling/ScriptableObjects/OneShotAnimationState.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 71196214900c82c479db633f3d431a4c
