+++ The Pom Game/Assets/Scripts/Attributes/Health.cs
@@ -1,7 +1,9 @@
+using Pom.AnimationHandling;
 using Pom.UndoSystem;
 using System;
 using UnityEngine;
 using UnityEngine.Events;
+using UnityEngine.InputSystem;
 
 namespace Pom.Attributes
 {
@@ -14,9 +16,19 @@ namespace Pom.Attributes
         public UnityEvent<GameObject> onTakeDamage;
         public UnityEvent onDeath;
 
+        AnimationStateMachine animationStateMachine => GetComponent<AnimationStateMachine>();
+
         private void Start()
         {
             CurrentHealth = StartingHealth;
+        }
+
+        void Update()
+        {
+            if (Keyboard.current.spaceKey.wasPressedThisFrame)
+            {
+                TakeDamage(1, null);
+            }
         }
 
         public void TakeDamage(int damage, GameObject aggressor)
@@ -24,13 +36,24 @@ namespace Pom.Attributes
             if(CurrentHealth > 0)
                 CurrentHealth = Mathf.Max(CurrentHealth - damage, 0);
 
+            if(aggressor != null)
             onTakeDamage?.Invoke(aggressor);
 
             if(CurrentHealth == 0)
             {
-                onDeath?.Invoke();
-                gameObject.SetActive(false);
+                Die();
             }
+            else
+            {
+                animationStateMachine.SwitchState(AnimationTag.Hit);
+            }
+        }
+
+        private void Die()
+        {
+            animationStateMachine.SwitchState(AnimationTag.Death);
+            animationStateMachine.onCurrentAnimationFinished += () => { gameObject.SetActive(false); };
+            onDeath?.Invoke();
         }
 
         public object CaptureState()
